Move RegionEq to separate file
